Some dynamic light sources optimizations for large mesh sizes. Added code to skip dlights texture update when there are no dlights this frame or the last frame and the textures are valid. -FG

+++ src/lightmap.h
@@ -95,7 +95,7 @@ public:
 class dls_cell {
 
 	float z1, z2;
-	vector<unsigned> lsrc;
+	vector<unsigned short> lsrc;
 
 public:
 	dls_cell() : z1(FAR_CLIP), z2(-FAR_CLIP) {}
@@ -109,7 +109,7 @@ public:
 	bool check_z(float z)    const {return (!empty() && z >= z1 && z <= z2);}
 	bool check_z_range(float zlo, float zhi) const {return (!empty() && zlo <= z2 && zhi >= z1);}
 	void get_close_sources(point const &pos, float radius, vector<unsigned> &dlights) const;
-	vector<unsigned> const &get_src_ixs() const {return lsrc;}
+	vector<unsigned short> const &get_src_ixs() const {return lsrc;}
 };
